Move `DateRange` type to src root, to make it globally available
