+++ ui/src/components/timetables/substitute-day-settings/SubstituteDaySettingsPage.tsx
@@ -10,12 +10,12 @@ import {
 import { Container, Row } from '../../../layoutComponents';
 import { selectTimetable } from '../../../redux';
 import { Path } from '../../../router/routeDetails';
+import { DateRange } from '../../../types';
 import { ConfirmationDialog } from '../../../uiComponents';
 import { CloseIconButton } from '../../../uiComponents/CloseIconButton';
 import { PageTitle } from '../../common';
 import { ObservationPeriodForm } from '../../forms/timetables/ObservationPeriodForm';
 import { CommonSubstitutePeriodSection } from './CommonSubstitutePeriod/CommonSubstitutePeriodSection';
-import { DateRange } from './DateRange';
 import { OccasionalSubstitutePeriodSection } from './OccasionalSubstitutePeriod';
 
 const testIds = {
