+++ ui/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './EnrichedStopRegistryTypes';
+export * from './DateRange';
 export * from './FilterableStopInfo';
 export * from './GeoHelpers';
 export * from './KeyHelpers';
